+++ libraries/management/commands/import_first_release_dates.py
@@ -1,6 +1,6 @@
 import djclick as click
 
-from libraries.github import LibraryUpdater
+from libraries.updater import LibraryUpdater
 from core.githubhelper import GithubAPIClient
 from libraries.models import Library
