+++ libraries/management/commands/import_library_versions.py
@@ -1,6 +1,6 @@
 import djclick as click
 
-from libraries.github import LibraryUpdater
+from libraries.updater import LibraryUpdater
 from core.githubhelper import GithubAPIClient, GithubDataParser
 from libraries.models import Library, LibraryVersion
 from libraries.tasks import get_and_store_library_version_documentation_urls_for_version
